добавил флаги, координаты в события и валидацию чтобы суда не заплывали на сушу
