+++ tracking/helpers/get_location_type.py
@@ -1,0 +1,7 @@
+from geopy import Nominatim
+
+
+def get_location_type(lat, lon):
+    geolocator = Nominatim(user_agent="location_tracker")
+    location = geolocator.reverse(f"{lat}, {lon}", exactly_one=True, addressdetails=True)
+    return location
