+++ tracking/serializers/ship.py
@@ -2,14 +2,21 @@ from rest_framework import serializers
 
 from tracking.models.ship import Ship
 
+FLAGS_STATIC_DIR = '/static/flags/'
+
 
 class ShipSerializer(serializers.ModelSerializer):
     country = serializers.SerializerMethodField()
     type = serializers.SerializerMethodField()
+    flag = serializers.SerializerMethodField()
 
     @staticmethod
     def get_country(obj):
         return obj.country.name
+
+    @staticmethod
+    def get_flag(obj):
+        return f"{FLAGS_STATIC_DIR}{obj.country.code.lower()}.gif"
 
     @staticmethod
     def get_type(obj):
@@ -17,4 +24,4 @@ class ShipSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = Ship
-        fields = ['id', 'country', 'code', 'tonnage', 'type', 'latitude', 'longitude']
+        fields = ['id', 'country', 'flag', 'code', 'tonnage', 'type', 'latitude', 'longitude']
